% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Add number}
\usage{
add(x, y)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
the sum of \code{x} and \code{y}
}
\description{
Add together two number
}
\details{
add @keywords
}
\examples{
add(1,1)
add(10,1)
}
