% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstpk.R
\docType{package}
\name{firstpk}
\alias{firstpk}
\alias{firstpk-package}
\title{The test packages provides some functions :}
\description{
The test packages provides some functions :
}
\section{test Functions}{

The foo function ...
}

